refactor output_df and run methods into parent bandit class, seperate testbed into a its own submodule
